rename bin files for clarify. init swap cleanup
